[refactor] Traffic enum 분리 #79
